requirements.txt
Flask
azure-storage-blob
azure-search-documents
openai
PyPDF2
